simplify data generation for least squares
